requirements.txt
streamlit>=1.28.0
pinecone-client>=3.0.0
openai>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-pinecone>=0.1.0
langchain-text-splitters>=0.2.0
langchain-core>=0.2.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
tiktoken>=0.5.0
volcengine-python-sdk>=1.0.0
